+++ src/com/web/oa/pojo/BaoxiaoBillExample.java
@@ -496,7 +496,7 @@ public class BaoxiaoBillExample {
             return (Criteria) this;
         }
 
-        public Criteria andUserIdEqualTo(Integer value) {
+        public Criteria andUserIdEqualTo(Long value) {
             addCriterion("user_id =", value, "userId");
             return (Criteria) this;
         }
